fix: resolve lint issues
